+++ lib/generators/happy_seed/roles/templates/ability.rb
@@ -1,0 +1,9 @@
+class Ability
+  include CanCan::Ability
+
+  def initialize(user)
+    if user && user.admin?
+      can :manage, :all
+    end
+  end
+end